Fix: extend Container component style in the correct way
When extending a react styled component, you have
to pass a "className" prop to the parent component.

When a component is extended via styled(component),
it creates a new css class with the new properties,
and applies that along with the css class applied
by the parent component - so you need to take in
the className on the parent component.

Missed the one line in the documentation that
explained this!