Ensure default product filters are run upon first render
Call useEffect to run filter on the first render.
Feels pretty tacky and lame.

If I am using useEffect to run the filter on initial
render, why not put query in the dependency array
and have it run every time it changes, instead of
passing callback functions to the subcomponents to
call when an input is changed?

UseEffect should be used for stuff external to react,
not internal stuff, so this just feels wrong. Probably,
my data flow/design sucks, so I will take another look
at it later.