Fix incorrect positioning of navBasket item counter
Using positioning properties like top, left, etc.
leaves the element in the wrong position if it
is added to the page after initial mount (e.g.
going from zero basket items, when no counter will
be displayed, to having more than zero, when the
counter will appear).

Commenting out "position: absolute", saving, then
uncommenting it and saving again also temporarily
fixes the issue.

I can only guess that the element gets confused
about where "top" is? Or what it is supposed to be
absolutely positioned against?

Checked the css generated by styled-components;
whether the counter is positioned correctly or not,
the css is exactly the same. So it is not an issue
with the css itself, but... something else?

Positioning with margins seems to work consistently,
so have replaced top/left properties with margin-top,
margin-left, etc.