Add comment explaining I am aware my code is terrible
But currently not sure how to fix it. In the interest
of maintaining forward momentum, settle for what I've got
and hopefully learn more and come back to this later.

Seems like this code is going against how react wants
to work, but I can't see any other way of doing it.
Probably as a consequence of a badly designed API or
application.