NavBasket: implement hacky solution for "forgotten" css styles
I wanted to absolutely position the number that
shows the number of items in the user's basket.
However, when the layout changes (e.g. adapts to
changing screen widths), the positioning gets
"forgotten" until the page is refreshed. I could
not figure out why this is.

Also trying to position the number was very wonky.
Again, not sure why the enclosing link didn't
expand to match its content. E.g. absolute positi-
oning to bottom: 0 did not align the number with
the bottom of its containing element.

Anyway, so did some dodgy hack with margins and
visibility. As a result, icon is slightly vertica-
lly misaligned.