NavBar: split desktop and mobile versions into their own components
The parent NavBar just controls which child component
is shown via css media queries. Nearly everything about
the styling and functionality is done via the subcomp-
onents.

The parent component passes NavLink and NavBasket to
the subcomponents. I wanted to just pass link data to
them instead, but this was complicated by having the
NavBasket component as a nav list item. It's not a
regular link, so instead I passed in JSX. So the nav
list items are styled in the parent component. Not
ideal. Will look at later and see if it can be changed
so the styling is entirely in the subcomponents, and
only the logic re. which subcomponent to show and hide
is in the parent NavBar component.

The code is a lot simpler and easier to follow.
This was a good change!