NavBar: move basket from hideable menu to always visible menu bar
Refactored styling of navLinks to be in its own file,
then used in the subcomponents, instead of passing in
an array of JSX as a prop.

So responsive layout is handled by navBar, and the
styling is entirely handled by the subcomponents,
desktopNavBar and mobileNavBar.

They both use navBarLink in order to share the styles,
but if the styles of the links need to diverge in the
future this would be very easy to do.