MobileProductFilter: simplify creation of aside element
MobileProductFilter started out the same as
DesktopProductFilter and was gradually modified
into the mobile version. As the desktop version
extended the ShopSectionContainer (which is a div),
it was passed "aside" as the "forwardedAs" prop so
it would be a complementary landmark instead of a
plain old div.

Since the mobile version later stopped extending
the ShopSectionContainer and just used a styled div,
this has been changed to a styled aside.