Delete redundant ProductFilter component
The functionality of this component (handling the
query, passing info down to the child filters,
choosing whether to display the mobile or desktop
version of the filters) has been raised up to the
Shop route.

This was in order to allow the mobile version of
the filter to use sticky positioning; the desktop
version of the filter is part of the main page
content, while the mobile version needed to be
outside of the main page content for sticky posit-
ioning to work properly.