Transition from getByText to getByTestId
The ShopSectionContainer tests were originally set
up to get the component by its text content, and
each test rendered with some text content just so
it could be grabbed. The component itself is just
a styled container - the only content would be
that which is provided in the children prop. There
are also no labels so I was not sure how to get
the rendered component when testing.

I learned about getByTestId and decided that would
be a better method. Getting by text content was
inadvertently testing whether children were inclu-
ded in the component (as the text content is a
child), even if the test was not about that.

There is a separate test to check whether the
elements in the children prop are included as
child nodes in the resulting component. Only this
test should be dependent on children being included,
not any others.

So this is a cleaner way of testing individual
pieces of functionality.